requirements.txt
django~=3.0.0
django-bootstrap4
django-debug-toolbar
djangorestframework
